refactor: relocate proof recognizers to a backend-neutral module
Move the shape recognizers (`RevOp` / `detect_rev_fn` /
`is_recursive_left_append` / `collect_rev_ops_in_law`, plus the generic AST
walkers `collect_called_fns` / `collect_called_fns_in_body` and `short_ctor`)
out of `dafny/` into a new `src/codegen/proof_recognize.rs`. They are pure
Aver-AST walks returning source-name structs — no backend syntax — so a single
module is the right home now that BOTH backends consume them: `dafny/lemmas.rs`
(Z3 renderer) and `lean/law_auto/induction.rs` (kernel renderer).

Removes the lingering lean→dafny import the previous commit left behind: the
Lean renderer no longer reaches into `codegen::dafny::lemmas`. `mod lemmas` is
private to `dafny` again.

Pure relocation, zero behaviour change: emitted Dafny is byte-identical across
the proof-corpus (SHA-verified), the rev tasks still kernel-prove on Lean, and
proof_spec is 90/90. Net -189 lines (the move also collapsed the cross-module
`super::toplevel::` indirection).
