+++ src/codegen/dafny/mod.rs
@@ -6,7 +6,7 @@
 /// file holding the trust header, top-level items, and verify lemmas.
 mod expr;
 mod fuel;
-pub(crate) mod lemmas;
+mod lemmas;
 mod toplevel;
 
 use crate::ast::{FnDef, TopLevel, VerifyKind};
